chore(website): split oversized test files below S1 gate, fix S3 hardcoded hosts [T001350]
Follow-up to the coverage-increase commit: the initial commit pushed
four test files (native-billing.test.ts, systemtest/cleanup.test.ts,
systemtest/failure-bridge.test.ts, website-db.test.ts) over the S1
600-line-per-file CI gate, which blocked the push. Splits each into
2-3 sibling files by describe-block/feature area (each fully
self-contained with its own imports/mocks), preserving every test
verbatim — no assertions changed, same total test count, same
per-file coverage % as before the split.

Also fixes 2 S3 (hardcoded-prod-host) violations in
systemtest/failure-bridge.test.ts and systemtest/test-run-bridge.test.ts:
replaced literal PROD_DOMAIN test values ('mentolder.de'/'korczewski.de')
with a synthetic 'example.test' domain — same code path under test
(URL construction from the PROD_DOMAIN env var), no real brand domain
hardcoded in test source.
