+++ docs/code-quality/repo-index.json
@@ -574,7 +574,7 @@
       "id": "website",
       "name": "Astro + Svelte website",
       "owner_agent": "bachelorprojekt-website",
-      "file_count": 1655,
+      "file_count": 1670,
       "files": [
         "website/.build-trigger",
         "website/.design-sync/NOTES.md",
@@ -1174,6 +1174,7 @@
         "website/src/lib/agentGuideSearch.ts",
         "website/src/lib/ai-metrics.test.ts",
         "website/src/lib/ai-metrics.ts",
+        "website/src/lib/appointments-db.test.ts",
         "website/src/lib/appointments-db.ts",
         "website/src/lib/assistant/actions-email.test.ts",
         "website/src/lib/assistant/actions.test.ts",
@@ -1182,6 +1183,7 @@
         "website/src/lib/assistant/actions/admin/finalizeMeeting.ts",
         "website/src/lib/assistant/actions/admin/index.ts",
         "website/src/lib/assistant/actions/admin/resolveTicket.ts",
+        "website/src/lib/assistant/actions/admin/scheduleFollowup.test.ts",
         "website/src/lib/assistant/actions/admin/scheduleFollowup.ts",
         "website/src/lib/assistant/actions/admin/sendInvoice.ts",
         "website/src/lib/assistant/actions/admin/writeClientNote.ts",
@@ -1229,6 +1231,7 @@
         "website/src/lib/billing-archive.ts",
         "website/src/lib/billing-audit.test.ts",
         "website/src/lib/billing-audit.ts",
+        "website/src/lib/billing-db.test.ts",
         "website/src/lib/billing-db.ts",
         "website/src/lib/billing-tax.test.ts",
         "website/src/lib/billing-tax.ts",
@@ -1238,6 +1241,7 @@
         "website/src/lib/bulk-status.test.ts",
         "website/src/lib/bulk-status.ts",
         "website/src/lib/caldav-cache.ts",
+        "website/src/lib/caldav.test.ts",
         "website/src/lib/caldav.ts",
         "website/src/lib/chunking.test.ts",
         "website/src/lib/chunking.ts",
@@ -1284,6 +1288,7 @@
         "website/src/lib/content-projection.ts",
         "website/src/lib/content-registry.test.ts",
         "website/src/lib/content-registry.ts",
+        "website/src/lib/content.test.ts",
         "website/src/lib/content.ts",
         "website/src/lib/cors.test.ts",
         "website/src/lib/cors.ts",
@@ -1297,6 +1302,7 @@
         "website/src/lib/db/filters.ts",
         "website/src/lib/delivery-metrics.test.ts",
         "website/src/lib/delivery-metrics.ts",
+        "website/src/lib/documents-db.test.ts",
         "website/src/lib/documents-db.ts",
         "website/src/lib/dora-metrics.test.ts",
         "website/src/lib/dora-metrics.ts",
@@ -1354,6 +1360,7 @@
         "website/src/lib/generation-jobs.ts",
         "website/src/lib/github-ci.test.ts",
         "website/src/lib/github-ci.ts",
+        "website/src/lib/goals-data.test.ts",
         "website/src/lib/goals-data.ts",
         "website/src/lib/graph-utils.test.ts",
         "website/src/lib/graph-utils.ts",
@@ -1427,7 +1434,9 @@
         "website/src/lib/markdown.ts",
         "website/src/lib/mediaviewer-bridge.test.ts",
         "website/src/lib/mediaviewer-bridge.ts",
+        "website/src/lib/meetings-db.test.ts",
         "website/src/lib/meetings-db.ts",
+        "website/src/lib/messaging-db-attachments.test.ts",
         "website/src/lib/messaging-db-attachments.ts",
         "website/src/lib/messaging-db-pool.ts",
         "website/src/lib/messaging-db.test.ts",
@@ -1445,6 +1454,7 @@
         "website/src/lib/nextcloud-files.ts",
         "website/src/lib/nextcloud-shares.test.ts",
         "website/src/lib/nextcloud-shares.ts",
+        "website/src/lib/nextcloud-talk-db.test.ts",
         "website/src/lib/nextcloud-talk-db.ts",
         "website/src/lib/notifications.test.ts",
         "website/src/lib/notifications.ts",
@@ -1456,6 +1466,7 @@
         "website/src/lib/pdf-a3-embed.ts",
         "website/src/lib/planning-office-types.ts",
         "website/src/lib/planning-office.clarify.test.ts",
+        "website/src/lib/planning-office.test.ts",
         "website/src/lib/planning-office.ts",
         "website/src/lib/platform-db.ensure.test.ts",
         "website/src/lib/platform-db.ts",
@@ -1554,15 +1565,19 @@
         "website/src/lib/systemtest/purge-all.ts",
         "website/src/lib/systemtest/reconciler.test.ts",
         "website/src/lib/systemtest/reconciler.ts",
+        "website/src/lib/systemtest/recorder.test.ts",
         "website/src/lib/systemtest/recorder.ts",
         "website/src/lib/systemtest/retest-trigger.test.ts",
         "website/src/lib/systemtest/seed-context.ts",
+        "website/src/lib/systemtest/test-run-bridge.test.ts",
         "website/src/lib/systemtest/test-run-bridge.ts",
         "website/src/lib/talk.test.ts",
         "website/src/lib/talk.ts",
         "website/src/lib/tax-monitor.test.ts",
         "website/src/lib/tax-monitor.ts",
+        "website/src/lib/test-infra-db.test.ts",
         "website/src/lib/test-infra-db.ts",
+        "website/src/lib/test-runner.test.ts",
         "website/src/lib/test-runner.ts",
         "website/src/lib/ticket-graph.test.ts",
         "website/src/lib/ticket-graph.ts",
